Failed attempt to make my program faster
I thought if the program would find the first reaction, then figure out
how big the "chain reaction" could be, then delete all of the rows from
*that*, then it would be faster.

For example, if the pattern was YAbCcBaZ, then the first reaction would
be Cc, but the chain reaction would be AbCcBa, leaving us with YZ.

Unfortunately, that didn't happen.  My original program took about 8 1/2
minutes after the data loading stage, and I killed this new and
unimproved version after it ran for 25 minutes with no answer.  Albeit,
the new version ran while my laptop was unplugged, so maybe there was
some power management happening, but still, not enough improvement, if
any, to make it worthwhile.

This new method *did* work on the sample input, for what it's worth.