Day 5 part 1 now takes 1 minute 7 seconds
with laptop plugged in.  Uses a completely different method.  The key is
that when I react two units (a and A, for example), then I should start
looking for the next reaction at the unit just before the two that I
reacted (removed), instead of going all the way back to the beginning of
the polymer string.

Still, I've read that others get the answer in under a second, perhaps
including part 2, so I'm sure there's more room for improvement.